+++ client/src/App.vue
@@ -221,7 +221,7 @@
 
 <script setup>
 import { ref, onMounted, onUnmounted } from 'vue';
-import GameApiService from './services/gameApi.js';
+import GameApiService from './services/gameApi';
 import LoginScreen from './components/LoginScreen.vue';
 import Lobby from './components/Lobby.vue';
 import GameScreen from './components/GameScreen.vue';
@@ -231,7 +231,7 @@ import ImprintPage from './components/ImprintPage.vue';
 import TrainingMode from './components/TrainingMode.vue';
 import CookieDisclaimer from './components/CookieDisclaimer.vue';
 import AppDialog from './components/AppDialog.vue';
-import analytics from './services/analytics.js';
+import analytics from './services/analytics';
 
 // Reactive state
 const currentUser = ref(null);
