+++ client/src/main.js
@@ -2,7 +2,7 @@ import { createApp } from 'vue';
 import './style.css';
 import './assets/fonts.css';
 import App from './App.vue';
-import analytics from './services/analytics.js';
+import analytics from './services/analytics';
 
 // Make analytics globally available
 window.analytics = analytics;
