+++ client/src/services/wordValidationService.ts
@@ -1,26 +1,22 @@
 import { germanWords, wordValidationUtils } from '../data/germanWords.js';
-import type { WordValidation } from '@/types';
 
 // Wort-Validierungs-Service für deutsche Wörter (Offline)
 class WordValidationService {
-  private cache: Map<string, WordValidation> = new Map(); // Cache für bereits validierte Wörter
-  private isEnabled: boolean = true; // Kann deaktiviert werden
-  private germanWords: Set<string> = germanWords; // Große deutsche Wortliste
+  constructor() {
+    this.cache = new Map(); // Cache für bereits validierte Wörter
+    this.isEnabled = true; // Kann deaktiviert werden
+    this.germanWords = germanWords; // Große deutsche Wortliste
+  }
 
   // Service aktivieren/deaktivieren
-  setEnabled(enabled: boolean): void {
+  setEnabled(enabled) {
     this.isEnabled = enabled;
   }
 
   // Wort validieren (Offline)
-  async validateWord(word: string): Promise<WordValidation> {
+  async validateWord(word) {
     if (!this.isEnabled) {
-      return {
-        isValid: true,
-        reason: 'Validation disabled',
-        word: word.trim().toLowerCase(),
-        source: 'offline',
-      };
+      return { isValid: true, reason: 'Validation disabled' };
     }
 
     const cleanWord = word.trim().toLowerCase();
@@ -30,30 +26,23 @@ class WordValidationService {
       return {
         isValid: false,
         reason: 'Wort zu kurz (mindestens 2 Buchstaben)',
-        word: cleanWord,
-        source: 'offline',
       };
     }
 
     // Nur Buchstaben erlauben
     if (!/^[a-zäöüß]+$/i.test(cleanWord)) {
-      return {
-        isValid: false,
-        reason: 'Nur Buchstaben erlaubt',
-        word: cleanWord,
-        source: 'offline',
-      };
+      return { isValid: false, reason: 'Nur Buchstaben erlaubt' };
     }
 
     // Cache prüfen
     if (this.cache.has(cleanWord)) {
-      return this.cache.get(cleanWord)!;
+      return this.cache.get(cleanWord);
     }
 
     // Offline-Validierung mit großer deutscher Wortliste
     const isValid = this.germanWords.has(cleanWord);
 
-    const result: WordValidation = {
+    const result = {
       isValid,
       reason: isValid
         ? 'Gültiges deutsches Wort'
@@ -67,10 +56,8 @@ class WordValidationService {
   }
 
   // Mehrere Wörter validieren
-  async validateWords(
-    words: string[]
-  ): Promise<Array<WordValidation & { word: string }>> {
-    const results: Array<WordValidation & { word: string }> = [];
+  async validateWords(words) {
+    const results = [];
     for (const word of words) {
       const result = await this.validateWord(word);
       results.push({ word, ...result });
@@ -79,12 +66,12 @@ class WordValidationService {
   }
 
   // Cache leeren
-  clearCache(): void {
+  clearCache() {
     this.cache.clear();
   }
 
   // Cache-Statistiken
-  getCacheStats(): { size: number; entries: Array<[string, WordValidation]> } {
+  getCacheStats() {
     return {
       size: this.cache.size,
       entries: Array.from(this.cache.entries()),
@@ -92,12 +79,12 @@ class WordValidationService {
   }
 
   // Wortliste-Statistiken
-  getWordListStats(): any {
+  getWordListStats() {
     return wordValidationUtils.getWordStats();
   }
 
   // Zufällige Wörter für Tests
-  getRandomWords(count: number = 10): string[] {
+  getRandomWords(count = 10) {
     return wordValidationUtils.getRandomWords(count);
   }
 }
